part3/requirements.txt
Flask==3.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
PyMySQL==1.1.1
pytest